Refactor: Extract components from large files for better maintainability
Split TestScoreView (553→460 lines) and TaskForm (471→360 lines) into smaller, focused components:

Created new components:
- ScoreCard.jsx: Test score card display (110 lines)
- CustomUnitForm.jsx: Custom unit creation form (60 lines)
- PastPaperFields.jsx: Past paper additional fields (70 lines)

Benefits:
- Improved code readability and maintainability
- Better separation of concerns
- Easier to test individual components
- Reduced component complexity

All functionality preserved with no breaking changes.
